refactor code practice one
