+++ practice-one/index.html
@@ -7,8 +7,8 @@
       type="image/x-icon"
       href="./src/assets/image/favicon.ico" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+    <meta name="description" content="Food ordering app" />
     <title>Shushi Restaurant</title>
-    <link rel="stylesheet" href="./index.css" />
   </head>
   <body>
     <div id="root" class="container"></div>
