+++ practice-one/src/App.tsx
@@ -1,18 +1,21 @@
-import { BrowserRouter, Route, Routes } from "react-router-dom";
 import "./App.css";
-import Menu from "./components/menu/Menu";
-import HomePage from "./components/home-page/HomePage";
-import Cart from "./components/cart/Cart";
-import Checkout from "./components/common/checkout/Checkout";
+
+import { BrowserRouter, Route, Routes } from "react-router-dom";
+
+import { Home } from "./components/index";
+import { routers } from "./Routers/Routers";
+import MainLayout from "./layout/MainLayout";
 
 function App() {
   return (
     <BrowserRouter>
       <Routes>
-        <Route path="/home" element={<HomePage />} />
-        <Route path="/menu" element={<Menu />} />
-        <Route path="/cart" element={<Cart />} />
-        <Route path="/checkout" element={<Checkout />} />
+        <Route path="/home" element={<Home />} />
+        <Route element={<MainLayout />}>
+          {routers.map((item) => (
+            <Route key={item.path} path={item.path} element={item.element} />
+          ))}
+        </Route>
       </Routes>
     </BrowserRouter>
   );
