+++ practice-one/src/Page/Checkout/index.tsx
@@ -1,0 +1,13 @@
+// Component
+import { FormCheckout } from "../../components";
+
+// Component Checkout
+const Checkout = () => {
+  return (
+    <>
+      <FormCheckout />
+    </>
+  );
+};
+
+export default Checkout;
