+++ practice-one/src/Page/Menu/index.tsx
@@ -1,25 +1,19 @@
 // CSS
-import "./listProduct.css";
-
-//React Hooks
-import React, { useEffect, useState } from "react";
+import "./Menu.css";
+// React Hook
+import React, { useContext, useEffect, useState } from "react";
 
 // Component
-import Image from "../common/image/Image";
-import Footer from "../footer/Footer";
-import Button from "../common/button/Button";
+import { ListProduct } from "../../components";
 
-// Interface
-import { CustomProductProps, ListProductProps } from "../../types/interface";
-
-// Constants, Image, Message error
-import { spinner } from "../../assets/image";
-import { BASE_URL } from "../../constants/base-url";
-import { ERROR_MESSAGES } from "../../constants/error";
+// LocalStorage
 import { getListCart } from "../../helper/data-localStorage";
+import { CustomProductProps } from "../../types/interface";
+import { BASE_URL } from "../../constants/base-url";
+import { CartLength } from "../../layout/MainLayout";
 
-// Component ListProduct
-const ListProduct: React.FC<ListProductProps> = ({ setCartLength }) => {
+// Component Menu
+const Menu: React.FC = () => {
   const [isLoading, setIsLoading] = useState(true);
   const [products, setProducts] = useState<CustomProductProps[]>([]);
   const [defaultValue, setDefaultValue] = useState(8);
@@ -39,6 +33,13 @@ const ListProduct: React.FC<ListProductProps> = ({ setCartLength }) => {
     localStorage.setItem("CartProducts", JSON.stringify(carts));
   }, [carts]);
 
+  const cartContext = useContext(CartLength);
+  if (cartContext === null) {
+    return null;
+  }
+
+  const { setCartLength } = cartContext;
+
   // Method Loadmore
   const showMorePoducts = () => {
     const newValue = defaultValue + 4;
@@ -55,7 +56,6 @@ const ListProduct: React.FC<ListProductProps> = ({ setCartLength }) => {
     product.isExist = true;
     setCarts((prevCart) => {
       const newCart = [...prevCart, product];
-      // Update the length of the shopping cart
       setCartLength(newCart.length);
       return newCart;
     });
@@ -102,56 +102,21 @@ const ListProduct: React.FC<ListProductProps> = ({ setCartLength }) => {
       setIsLoading(false);
     }
   };
-
-  // If there is an error, display the error message
-  if (error) {
-    return (
-      <p className="messages-error text-medium">{ERROR_MESSAGES.fetchError}</p>
-    );
-  }
-
-  // If loading data, display the loading icon
-  if (isLoading) {
-    return (
-      <div className="spinner">
-        <Image src={spinner} />
-      </div>
-    );
-  }
-
   return (
-    <section className="section-menu font-family">
-      <h3 className="text-h3">Sushi food</h3>
-      <hr className="dash dash-menu"></hr>
-      <ul className="list-menu">
-        {products.slice(0, defaultValue).map((product) => {
-          return (
-            <li className="menu-item" key={product.id}>
-              <Image
-                className={` img-rectangle ${
-                  isInCart(product.id) ? "added-to-cart" : ""
-                }`}
-                src={`${product.image}`}
-                alt={product.name}
-                onClick={() => {
-                  isInCart(product.id)
-                    ? removeFromCart(product.id)
-                    : addToCart(product);
-                }}
-              />
-              <span className="text-small">{product.name}</span>
-            </li>
-          );
-        })}
-      </ul>
-      <Button
-        textBtn="Load more "
-        className={`btn-item secondary-text-btn ${isFull ? "isFull" : ""}`}
-        onClick={showMorePoducts}
+    <>
+      <ListProduct
+        addToCart={addToCart}
+        removeFromCart={removeFromCart}
+        isInCart={isInCart}
+        showMorePoducts={showMorePoducts}
+        isFull={isFull}
+        defaultValue={defaultValue}
+        error={error}
+        products={products}
+        isLoading={isLoading}
       />
-      <Footer className="copyright-text" />
-    </section>
+    </>
   );
 };
 
-export default ListProduct;
+export default Menu;
