+++ practice-one/src/Page/Menu/menu.css
@@ -3,13 +3,9 @@
   margin: auto;
 }
 
-
-
 .dash-menu {
   width: 306px;
 }
-
-
 
 .copyright {
   margin-top: 208px;
