+++ practice-one/src/Routers/Routers.tsx
@@ -1,0 +1,16 @@
+import { Cart, Checkout, Menu } from "../components";
+
+export const routers = [
+  {
+    path: "/menu",
+    element: <Menu />,
+  },
+  {
+    path: "/cart",
+    element: <Cart />,
+  },
+  {
+    path: "/checkout",
+    element: <Checkout />,
+  },
+];
