--- practice-one/src/components/checkout/Checkout.tsx
@@ -1,15 +1,0 @@
-// Component
-import Navbar from "../common/navbar/Navbar";
-import FormCheckout from "../form-checkout/FormCheckout";
-
-// Component Checkout
-const Checkout = () => {
-  return (
-    <section>
-      <Navbar />
-      <FormCheckout />
-    </section>
-  );
-};
-
-export default Checkout;
