+++ practice-one/src/components/common/button/Button.stories.tsx
@@ -1,5 +1,5 @@
 import type { Meta, StoryObj } from "@storybook/react";
-import Button from "./Button";
+import Button from ".";
 
 export default {
   title: "Components/Button",
