+++ practice-one/src/components/common/card/Card.stories.tsx
@@ -1,5 +1,5 @@
 import { Meta, StoryObj } from "@storybook/react";
-import Card from "./Card";
+import Card from ".";
 
 export default {
   title: "Components/Card",
