--- practice-one/src/components/common/checkout/Checkout.tsx
@@ -1,15 +1,0 @@
-// Component
-import FormCheckout from "../../form-checkout/FormCheckout";
-import Navbar from "../navbar/Navbar";
-
-// Component Checkout
-const Checkout = () => {
-  return (
-    <section className="component-layout font-family">
-      <Navbar />
-      <FormCheckout />
-    </section>
-  );
-};
-
-export default Checkout;
