+++ practice-one/src/components/common/index.tsx
@@ -1,0 +1,6 @@
+export { default as Button } from "./Button/index";
+export { default as Card } from "./Card/index";
+export { default as Image } from "./Image/index";
+export { default as Input } from "./Input/index";
+export { default as Label } from "./Label/index";
+export { default as Navbar } from "./Navbar/index";
