+++ practice-one/src/components/common/input/Input.stories.tsx
@@ -1,5 +1,5 @@
 import { Meta, StoryObj } from "@storybook/react";
-import Input from "./Input";
+import Input from ".";
 
 export default {
   title: "Components/Input",
