+++ practice-one/src/components/footer/index.tsx
