+++ practice-one/src/components/index.tsx
@@ -1,0 +1,10 @@
+export { Button, Card, Image, Input, Label, Navbar } from "./common/index";
+
+export { default as Cart } from "../Page/Cart/index";
+export { default as Checkout } from "../Page/Checkout/index";
+export { default as Footer } from "./Footer/index";
+export { default as FormCheckout } from "./FormCheckout/index";
+export { default as Home } from "../Page/Home/index";
+export { default as ListCart } from "./ListCart/index";
+export { default as ListProduct } from "./ListProduct/index";
+export { default as Menu } from "../Page/Menu/index";
